# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Febrafar - Planner",
        "description": "Febrafar planner",
        "version": "1.0"
    },
    "paths": {
...
        "/api/task-types/matches": {
            "get": {
                "tags": [
                    "Task Types"
                ],
                "summary": "Find all task types that match certain criteria",
                "operationId": "1bdf8c397dfac9b4559e23126cfa9e03",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name to match",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status to match",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of matching task types successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TaskType"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/task-types": {
            "get": {
                "tags": [
                    "Task Types"
                ],
                "summary": "Retrieve the list of task types",
                "operationId": "c0f6a658288861095bc9bf46b3790a42",
                "responses": {
                    "200": {
                        "description": "List of all task types successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TaskType"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Task Types"
                ],
                "summary": "Create a new task type",
                "operationId": "ac5bd38d7271502c0b02c6da374d6d81",
                "requestBody": {
                    "description": "Task type creation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the new task type",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status of the new task type",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task type successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskType"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Invalid input data"
                    }
                }
            }
        },
        "/api/task-types/{type}": {
            "get": {
                "tags": [
                    "Task Types"
                ],
                "summary": "Retrieve a single task type",
                "operationId": "838e18c3d9fa6cccfc3ef8e225253dac",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "ID of the task type to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task type successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskType"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Task type not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Task Types"
                ],
                "summary": "Update an existing task type",
                "operationId": "9de535bdd237629588f98db7532afba2",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "ID of the task type to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated task type data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Updated name of the task type",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Updated status of the task type",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task type successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskType"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Task type not found"
                    },
                    "422": {
                        "description": "Invalid input data"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Task Types"
                ],
                "summary": "Delete an existing task type",
                "operationId": "b33db54c3213b3eb497f2cbdb6cb6d56",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "ID of the task type to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Task type successfully deleted"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Task type not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TaskType": {
                "properties": {
                    "id": {
                        "description": "Task Type ID",
                        "type": "string"
                    },
                    "name": {
                        "description": "Task Type name",
                        "type": "string"
                    },
                    "status": {
                        "description": "Task Type status",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}